+++ pypeit/scripts/launch_identify.py
@@ -40,8 +40,9 @@ def main(args):
     import astropy.io.fits as fits
     from pypeit.masterframe import MasterFrame
     from pypeit.spectrographs.util import load_spectrograph
+    from pypeit.core import parse
     from pypeit.core.gui import identify as gui_identify
-    from pypeit.core.wavecal import waveio
+    from pypeit.core.wavecal import waveio, templates
     from pypeit.wavecalib import WaveCalib
     from pypeit import arcimage, edgetrace
     from pypeit.images import pypeitimage
@@ -84,15 +85,24 @@ def main(args):
     # Check if a solution exists
     solnname = os.path.join(mdir, MasterFrame.construct_file_name('WaveCalib', mkey, file_format='json'))
     wv_calib = waveio.load_wavelength_calibration(solnname) if os.path.exists(solnname) else None
-    # Select the appropriate detector
-    if wv_calib is not None:
-        wv_calib = wv_calib[str(args.det-1)]
 
     # Load the MasterFrame (if it exists and is desired)?
     wavecal = WaveCalib(msarc, tslits_dict, spec, par)
     arccen, arc_maskslit = wavecal.extract_arcs()
 
-    # Launch the identify window
-    arcfitter = gui_identify.initialise(arccen, slit=int(args.slit), par=par, wv_calib=wv_calib)
+    binspec, binspat = parse.parse_binning(spec.get_meta_value(msarc.head0['F1'], 'binning'))
 
-    # TODO :: Ask the user if they wish to store the result in PypeIt calibrations
+    # Launch the identify window
+    arcfitter = gui_identify.initialise(arccen, slit=int(args.slit), par=par, wv_calib_all=wv_calib)
+    final_fit = arcfitter.get_results()
+
+    # Ask the user if they wish to store the result in PypeIt calibrations
+    ans = ''
+    while ans != 'y' and ans != 'n':
+        ans = input("Would you like to store this wavelength solution in the archive? (y/n):")
+    if ans == 'y' and final_fit['rms'] < 0.1:
+        gratname = fits.getheader(msarc.head0['F1'])[spec.meta['dispname']['card']].replace("/", "_")
+        dispangl = "UNKNOWN"
+        templates.pypeit_identify_record(final_fit, binspec, specname, gratname, dispangl)
+        print("Your wavelength solution has been stored")
+        print("Please consider sending your solution to the PypeIt team!")